+++ NEWS.md
@@ -1,4 +1,5 @@
 # unheadr (development version)
+* Adds `annotate_mf_all()` to translate meaningful formatting for all cells in a spreadsheet.
 
 # unheadr 0.1.0